Move instrumentation classes into one package
